requirements.txt
flask==2.3.3
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
torch==2.0.1
transformers==4.33.1
python-dotenv==1.0.0
joblib==1.3.2
pytest==7.4.0
black==23.7.0
flake8==6.1.0
